pyproject.toml
[project]
name = "uv-django"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.2",
    "django-environ>=0.12.1",
    "djangorestframework>=3.16.1",
    "whitenoise>=6.11.0",
    "markdown>=3.7",
]

[dependency-groups]
dev = [
    "ruff>=0.15.1",
]
